+++ replacer-backend/src/test/java/es/bvalero/replacer/wikipedia/api/WikipediaApiServiceTest.java
@@ -104,8 +104,8 @@ class WikipediaApiServiceTest {
         when(wikipediaApiRequestHelper.executeApiRequest(any(WikipediaApiRequest.class))).thenReturn(response);
 
         Collection<WikipediaPage> pages = wikipediaService.getPagesByIds(
-            List.of(6219990, 6903884),
-            WikipediaLanguage.SPANISH
+            WikipediaLanguage.SPANISH,
+            List.of(6219990, 6903884)
         );
         assertNotNull(pages);
         assertEquals(2, pages.size());